requirements.txt
qrcode[pil]==7.4.2
django==5.0.4
djangorestframework==3.15.1
pytz==2024.1
django-admin-interface==0.28.6
django-simple-history==3.11.0
gunicorn==23.0.0
psycopg[binary]==3.2.12
openpyxl==3.1.5
djangorestframework_simplejwt==5.3.1
django-storages[s3]==1.14.4
boto3==1.35.99
drf-nested-routers==0.94.1
pyfcm==2.0.7
django-cors-headers==4.7.0
python-telegram-bot==20.8
django_apscheduler==0.7.0
apscheduler==3.11.2
workalendar==17.0.0
structlog==25.5.0
